Remove `--version` and simplify
Remove `globus --version` in favor of `globus version`. closes #421

To simplify, consolidate all of the version helpers into the version
command file, so that responsibility is spread over fewer
files/modules/subpackages.
